test(e2e): refactor game upload tests into separate file
